Renamed to GroupWithdrawal.tsx-> GenericModal.tsx and update for general use
