+++ src/components/todo/GroupMenuButton.tsx
@@ -2,16 +2,25 @@ import React from 'react';
 import Menu from '@material-ui/core/Menu';
 import MenuItem from '@material-ui/core/MenuItem';
 import MoreHorizIcon from '@material-ui/icons/MoreHoriz';
-import { EditGroupMembers, EditGroupName, GroupWithdrawal } from './index';
-import { Group } from '../../reducks/groups/types';
 import { IconButton } from '@material-ui/core';
+import { getApprovedGroups } from '../../reducks/groups/selectors';
+import { GenericModal } from '../uikit';
+import { EditGroupMembers, EditGroupName } from './index';
+import { Group } from '../../reducks/groups/types';
+import { useDispatch, useSelector } from 'react-redux';
+import { State } from '../../reducks/store/types';
+import { groupWithdrawal } from '../../reducks/groups/operations';
 
 interface MenuButtonProps {
   approvedGroup: Group;
 }
 
 const GroupMenuButton = (props: MenuButtonProps) => {
+  const dispatch = useDispatch();
   const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
+  const selector = useSelector((state: State) => state);
+  const approvedGroups = getApprovedGroups(selector);
+  const groupId = props.approvedGroup.group_id;
 
   const handleClick = (event: React.MouseEvent<HTMLButtonElement>) => {
     setAnchorEl(event.currentTarget);
@@ -20,6 +29,15 @@ const GroupMenuButton = (props: MenuButtonProps) => {
   const handleClose = () => {
     setAnchorEl(null);
   };
+
+  const approvedGroupIds = approvedGroups.map((approvedGroup) => {
+    return approvedGroup.group_id;
+  });
+
+  const nextGroupIds: number[] = approvedGroupIds.filter((approvedGroupId) => {
+    return approvedGroupId !== groupId;
+  });
+  const nextGroupId = nextGroupIds[0];
 
   return (
     <>
@@ -39,7 +57,12 @@ const GroupMenuButton = (props: MenuButtonProps) => {
           approvedGroup={props.approvedGroup}
         />
         <EditGroupMembers modalTitleLabel={'メンバーの編集'} approvedGroup={props.approvedGroup} />
-        <GroupWithdrawal approvedGroup={props.approvedGroup} />
+        <GenericModal
+          dispatch={() => dispatch(groupWithdrawal(groupId, nextGroupId))}
+          menuLabel={'グループを退会'}
+          modalText={`${props.approvedGroup.group_name}を退会しますか？`}
+          buttonLabel={'退会'}
+        />
         <MenuItem onClick={handleClose}>グループを削除</MenuItem>
       </Menu>
     </>
