+++ src/components/todo/index.tsx
@@ -4,7 +4,6 @@ export { default as EditGroupMembers } from './EditGroupMembers';
 export { default as EditGroupName } from './EditGroupName';
 export { default as GroupMenuButton } from './GroupMenuButton';
 export { default as GroupName } from './GroupName';
-export { default as GroupWithdrawal } from './GroupWithdrawal';
 export { default as InputTodoList } from './InputTodoList';
 export { default as InvitationNotifications } from './InvitationNotifications';
 export { default as MobileTodoMenu } from './MobileTodoMenu';
