+++ src/components/uikit/index.tsx
@@ -1,8 +1,9 @@
 export { default as AddButton } from './AddButton';
-export { default as GenericButton } from './GenericButton';
-export { default as DatePicker } from './DatePicker';
 export { default as CategoryInput } from './CategoryInput';
-export { default as TextInput } from './TextInput';
+export { default as DatePicker } from './DatePicker';
+export { default as GenericButton } from './GenericButton';
+export { default as GenericModal } from './GenericModal';
+export { default as InputModal } from './InputModal';
 export { default as KindSelectBox } from './KindSelectBox';
 export { default as ModalInform } from './ModalInform';
-export { default as InputModal } from './InputModal';
+export { default as TextInput } from './TextInput';
